Move BCTSN umbersData in file
