+++ src/libs/antares/study/scenario-builder/TSnumberData.h
@@ -265,46 +265,6 @@ inline CString<512, false> renewableTSNumberData::get_prefix() const
     return "r,";
 }
 
-// =================================
-// Binding Constraints ...
-// =================================
-
-class BindingConstraintsTSNumberData : public TSNumberData
-{
-public:
-    BindingConstraintsTSNumberData() = default;
-    virtual ~BindingConstraintsTSNumberData() = default;
-
-    bool reset(const Study& study) override;
-    void saveToINIFile(const Study& study, Yuni::IO::File::Stream& file) const override;
-
-    void setData(const std::string& group_name, unsigned year, unsigned value);
-    unsigned get(const std::string& group_name, unsigned year) const;
-    bool apply(Study& study) override;
-    CString<512, false> get_prefix() const override;
-    unsigned get_tsGenCount(const Study& study) const override;
-private:
-    std::map<std::string, MatrixType> rules_;
-};
-
-inline unsigned BindingConstraintsTSNumberData::get(const std::string& group_name, const unsigned year) const
-{
-    auto it = rules_.find(group_name);
-    if (it == rules_.end()) {
-        return 0;
-    }
-    return it->second[0][year];
-}
-
-inline CString<512, false> BindingConstraintsTSNumberData::get_prefix() const
-{
-    return "bc,";
-}
-
-inline unsigned BindingConstraintsTSNumberData::get_tsGenCount(const Study&) const {
-    return 0;
-}
-
 } // namespace Antares::Data::ScenarioBuilder
 
 #endif // __LIBS_STUDY_SCENARIO_BUILDER_DATA_TS_NUMBER_H__
