+++ src/libs/antares/study/scenario-builder/rules.h
@@ -33,6 +33,7 @@
 #include "TSnumberData.h"
 #include "hydroLevelsData.h"
 #include "NTCTSNumberData.h"
+#include "BindingConstraintsTSNumbersData.h"
 #include <map>
 #include <memory>
